# openapi/notebook-api.yaml
openapi: 3.1.0
info:
  title: Notebook API
  description: |
    REST API for managing notebooks.
    This OpenAPI document is the source of truth for the Notebook REST API.
    Java API interfaces and DTOs are generated from this contract.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.example.com
    description: Production
tags:
  - name: Notebooks
    description: Notebook management operations
paths:
  /api/v1/notebooks:
    get:
      tags:
        - Notebooks
      summary: Get all notebooks
      description: Returns all notebooks.
      operationId: getAllNotebooks
      responses:
        '200':
          description: Successfully retrieved notebooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notebook'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Notebooks
      summary: Create a notebook
      description: Creates a new notebook.
      operationId: createNotebook
      requestBody:
        required: true
        description: Notebook to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNotebookRequest'
      responses:
        '201':
          description: Notebook successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notebook'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/notebooks/{id}:
    get:
      tags:
        - Notebooks
      summary: Get notebook by ID
      operationId: getNotebookById
      parameters:
        - $ref: '#/components/parameters/NotebookId'
      responses:
        '200':
          description: Notebook found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notebook'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Notebooks
      summary: Update notebook
      operationId: updateNotebook
      parameters:
        - $ref: '#/components/parameters/NotebookId'
      requestBody:
        required: true
        description: Updated notebook
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNotebookRequest'
      responses:
        '200':
          description: Notebook successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notebook'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Notebooks
      summary: Delete notebook
      operationId: deleteNotebook
      parameters:
        - $ref: '#/components/parameters/NotebookId'
      responses:
        '204':
          description: Notebook successfully deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  parameters:
    NotebookId:
      name: id
      in: path
      required: true
      description: Unique identifier of the notebook
      schema:
        type: string
        format: uuid
  schemas:
    Notebook:
      type: object
      required:
        - id
        - title
        - description
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          description: Unique notebook identifier
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Notebook title
        description:
          type:
            - string
          maxLength: 5000
          description: Optional notebook description
        createdAt:
          type: string
          format: date-time
          readOnly: true
          description: Notebook creation timestamp
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          description: Last notebook modification timestamp
    CreateNotebookRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Notebook title
        description:
          type:
            - string
            - 'null'
          maxLength: 5000
          description: Optional notebook description
    UpdateNotebookRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Notebook title
        description:
          type:
            - string
            - 'null'
          maxLength: 5000
          description: Optional notebook description
    ProblemDetail:
      type: object
      required:
        - status
        - title
        - detail
        - timestamp
      properties:
        status:
          type: integer
          format: int32
          description: HTTP status code
        title:
          type: string
          description: Short description of the error
        detail:
          type: string
          description: Detailed error description
        instance:
          type: string
          description: Request URI
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
          description: Distributed tracing identifier
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: Notebook not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    InternalServerError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'